Use Kivy.uix.image in place of Kivy.uix.asyncimage
Originally we switched to using AsyncImage instead of Image because we
were running into segfaults when running the DD. While the AsyncImage
does seem to solve this issue it creates new bug in that the images do
not load.

The source of the segfault was due to on config change not all widgets
were being cleared/destroyed which lead to using more and more memory as
the config was reloaded.