Use same instance of digitaldash.pid objects for same pids
By using the same instance of our pid object we can ensure that things
like the max value observed for some pid are recorded across multiple
views.

For example if two views have RPM as a gauge, it makes sense that the
min/max value for both RPM gauges should be the same.